Revert "Merge branch 'develop/chethiya' into 'main'"
This reverts commit 111fc69ff45a6ca0cbae3ec8985b6a031a6b3261.

+++ lib/models/task/task.dart
@@ -9,7 +9,7 @@ class Task {
 
   String taskName;
 
-  double estimatedHours;
+  int estimatedHours;
 
   DateTime createdAt;
